requirements.txt
asgiref==3.8.1
Django==5.2.1
setuptools==65.5.1
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
uvicorn
sqlalchemy
flake8
flake8-docstrings
flake8-import-order
pytest
pytest-cov
httpx
bandit
pip-audit
uvicorn
sqlalchemy
pytest-django==4.5.2
python-dotenv==1.1.0
